requirements.txt
# Core dependencies for n8n workflow automation
openai>=1.0.0
python-dotenv>=0.19.0

# Image processing dependencies
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0

# Additional utilities (if needed for future development)
# requests>=2.25.0
# pandas>=1.3.0
